chore: extract CLI context into src/context.ts
Part of REAL-1037.
